# condition distribution to os_family

+++ server/init/roles/os/templates/mirror.sh
@@ -3,21 +3,18 @@ set -e
 
 os_distribution="{{ param_ansible_distribution }}"
 os_major_version="{{ param_ansible_distribution_major_version }}"
-os_product_name="{{ param_ansible_product_name }}"
 
 if [ "centos" = "${os_distribution}" ]; then
     case ${os_major_version} in
         "9")
             for r in `find /etc/yum.repos.d/ -maxdepth 1 -name "centos*.repo"`;do
               r_bak="${r}.bak"
-              if [ ! -f "${r_bak}" ];then
-                cp -rv ${r} ${r_bak}
-              fi
+              [ ! -f "${r_bak}" ] && cp -rv ${r} ${r_bak}
             done
             cp -r {{ param_remote_role_path }}/*.repo /etc/yum.repos.d/
             ;;
         *)
-          echo "nothing"
+          echo "unknown os_major_version"
     esac
 fi
 
@@ -26,20 +23,16 @@ if [ "rocky" = "${os_distribution}" ]; then
         "9")
             for r in `find /etc/yum.repos.d/ -maxdepth 1 -name "rocky*.repo"`;do
               r_bak="${r}.bak"
-              if [ ! -f "${r_bak}" ];then
-                cp -rv ${r} ${r_bak}
-              fi
+              [ ! -f "${r_bak}" ] && cp -rv ${r} ${r_bak}
               sed -e 's|^mirrorlist=|#mirrorlist=|g' -e 's|^#baseurl=http://dl.rockylinux.org/$contentdir|baseurl=http://{{ param_mirror_yum }}/rocky|g' ${r_bak} | tee ${r} > /dev/null
             done
             ;;
         *)
-          echo "nothing"
+          echo "unknown os_major_version"
     esac
 fi
 
 if [ "ubuntu" = "${os_distribution}" ]; then
-  if [ ! -f "/etc/apt/sources.list.bak" ];then
-    cp -rv /etc/apt/sources.list /etc/apt/sources.list.bak
-  fi
+  [ ! -f "/etc/apt/sources.list.bak" ] && cp -rv /etc/apt/sources.list /etc/apt/sources.list.bak
   sed "s,\(http\|https\)://.*.ubuntu.com,http://{{ param_mirror_apt }},g" /etc/apt/sources.list.bak | tee /etc/apt/sources.list > /dev/null
 fi
